% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeSNPDirectory.R
\name{AnalyzeSNPDirectory}
\alias{AnalyzeSNPDirectory}
\title{Does a rough, first pass SNP FISH analysis for a given directory}
\usage{
AnalyzeSNPDirectory(masterdir, outdir, SNPTableFlag = FALSE)
}
\arguments{
\item{masterdir}{directory to search for all sub yaml-files for standard
analysis}

\item{outdir}{directory in which to output SNP plots}

\item{SNPTableFlag}{Will produce plots on SNP tables if they exist. Default
is false.}
}
\description{
Does a rough, first pass SNP FISH analysis for a given directory
}

