% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFilesFromYamlGroup.R
\name{ExtractFilesFromYamlGroup}
\alias{ExtractFilesFromYamlGroup}
\title{Loads Yaml file and SNPTables from a YamlGroup}
\usage{
ExtractFilesFromYamlGroup(yaml.group, masterDir, guide.header = "_guide_",
  cy.header = "_cy_", tmr.header = "_tmr_")
}
\arguments{
\item{yaml.group}{the list of directories of a yaml file and its associated
SNP tables}

\item{masterDir}{directory where yamlGroup was called}

\item{guide.header}{tag to search for SNP guide table. The default is
"_guide_".}

\item{cy.header}{tag to search for tmr or SNP 1 table. The default is
"_tmr_"}

\item{tmr.header}{tag to search for cy or SNP 2 table. The default is
"_cy_"}
}
\value{
A list ordered in the following way:
 The first object will the the experimental data loaded from the Yaml file
 The second object will the guide SNP data table
 The third object will be the tmr SNP data table.
 The fourth object will be the cy SNP data table.
}
\description{
SNPTables are extracted from a YamlGroup according to the headers used in
the file names. Modify defaults as required. Objects not present will have
their places held by NA.
}

