% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllYamlGroupsinDir.R
\name{GetAllYamlGroupsinDir}
\alias{GetAllYamlGroupsinDir}
\title{Searches the given directory for all readMe.yaml files}
\usage{
GetAllYamlGroupsinDir(dir, snptable.header = "SNPTable_",
  yamlfile.header = "readMe.yaml")
}
\arguments{
\item{dir}{the directory that you are searching within}

\item{snptable.header}{the key text to grep when searching for the SNP csv.
Default is "SNPTable_".}

\item{yamlfile.header}{the key text to grep when searching for the readMe.yaml.
Default is "readMe.yaml".}
}
\value{
A list with each element corresponding to a unique readMe.yaml file. For
  each of those elements, the first entry will be that yaml file and
  subsequent entries will be the associated SNP csv files. Will return NA
  if there are no paired SNP csv files found with a given readMe.yaml file.
}
\description{
Searches the given directory for all readMe.yaml files within all
subdirectories as well as their associated files their associated
"SNPTable_" files in that directory.
  NOTE: This function assumes that your YAML readme is in the same
         directory as its paired SNP CSV data. If thatis not the case,
         it will not work correctly.
}

