% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggname.R
\name{ggname}
\alias{ggname}
\title{Make catagory names friendly to ggplot2}
\usage{
ggname(x)
}
\arguments{
\item{x}{string to make acceptable to ggplot2}
}
\value{
x string made acceptable to ggplot2
}
\description{
Changes a string so it is enclosed in back-ticks.
This can be used to make column names that have spaces (blanks)
or non-letter characters acceptable to ggplot2.
This version of the function is vectorized with sapply.
}
\examples{
 ggname("B6 Allele")
}

