% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{MakeAlleleScatterPlot}
\alias{MakeAlleleScatterPlot}
\title{Creates scatter plot between SNPA vs SNPB counts}
\usage{
MakeAlleleScatterPlot(exp.data, guide.table, outdir)
}
\arguments{
\item{exp.data}{experimental data from YAML file}

\item{guide.table}{Guide SNPTable to be ploted}

\item{outdir}{Directory to save output to}
}
\description{
Can be used to measure intrinsic vs extrinsic noise
}

