% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGuideTablePlots.R
\name{MakeGuideTablePlots}
\alias{MakeGuideTablePlots}
\title{Makes a standard set of graphs for guide table}
\usage{
MakeGuideTablePlots(exp.data, guide.table, outdir)
}
\arguments{
\item{guide.table}{SNP Table for guide probe}

\item{outdir}{directory in which to output SNP plots}

\item{exp.Data}{experiment data extracted from YAML file}
}
\description{
Makes a standard set of graphs for guide table
}

