% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{MakeOverallImbalanceBar}
\alias{MakeOverallImbalanceBar}
\title{Creates bar graph comparing ratio of SNP A to SNP B in cell population}
\usage{
MakeOverallImbalanceBar(exp.data, guide.table, outdir)
}
\arguments{
\item{exp.data}{experimental data from YAML file}

\item{guide.table}{Guide SNPTable to be ploted}

\item{outdir}{Directory to save output to}
}
\description{
Creates bar graph comparing ratio of SNP A to SNP B in cell population
}

