% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{MakeStackedCellCountPlot}
\alias{MakeStackedCellCountPlot}
\title{Creates standard bar plot of stacked cell counts}
\usage{
MakeStackedCellCountPlot(exp.data, snp.table, outdir)
}
\arguments{
\item{exp.data}{experimental data from YAML file}

\item{snp.table}{SNPTable to be ploted}

\item{outdir}{Directory to save output to}
}
\description{
Each bar is an individual cell. The number on top of each bar repsents the
percentage of detected transcripts in that cell, including 3-color spots.
}

