% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{MakeTotalCellCountPie}
\alias{MakeTotalCellCountPie}
\title{Creates standard pie chart for cell population average}
\usage{
MakeTotalCellCountPie(exp.data, snp.table, outdir)
}
\arguments{
\item{exp.data}{experimental data from YAML file}

\item{snp.table}{SNPTable to be ploted}

\item{outdir}{Directory to save output to}
}
\description{
Each slice represents the propoertion that was detected to the according label.
}

