% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReorderSNPLevels.R
\name{ReorderSNPLevels}
\alias{ReorderSNPLevels}
\title{Reorders the label's levels in table such that undetected spots and
three color are stacked on top of the SNP labels}
\usage{
ReorderSNPLevels(snp.table, undetec.label = "undetec",
  threecolor.label = "3-color")
}
\arguments{
\item{snp.table}{the standard SNP table whose levels will be reordered}

\item{undetec.label}{label for undetected spots. Default is "undetec".}

\item{threecolor.label}{label used for three-color spots. Default is "3-color".}
}
\value{
snp.table with levels reordered
}
\description{
Reorders the label's levels in table such that undetected spots and
three color are stacked on top of the SNP labels
}

